requirements.txt
Flask==2.0.1
web3==6.0.0
python-dotenv==0.19.0
flask-sqlalchemy==2.5.1
